feat: MCP OAuth security hardening and production readiness
Implemented security improvements and production configuration for MCP OAuth:

- Added constant-time secret comparison to prevent timing attacks
- Implemented token logging sanitization to protect sensitive credentials
- Fixed timestamp overflow (Integer → Long) to prevent 2038 issues
- Added input validation for connector names
- Implemented HttpClient resource cleanup (AutoCloseable)
- Added token refresh response validation with null checks
- Replaced hardcoded base URL with dynamic SystemRepository configuration
- Fixed MCP Inspector compatibility (removed unimplemented logging capability)
- Added example credential files and test setup documentation
- Removed commented code and unused files for cleaner codebase

Security TODOs documented for future work:
- Race condition in authorization code exchange (requires DB schema changes)
- Rate limiting for OAuth endpoints (requires new infrastructure)

Testing:
- All changes tested with Snowflake OAuth connector
- MCP Inspector connection verified working
- Code formatted with spotless

Breaking Changes: None

+++ openmetadata-mcp/src/main/java/org/openmetadata/mcp/server/auth/provider/SimpleAuthProvider.java
@@ -110,7 +110,7 @@ public class SimpleAuthProvider implements OAuthAuthorizationServerProvider {
     accessToken.setToken(accessTokenValue);
     accessToken.setClientId(client.getClientId());
     accessToken.setScopes(authorizationCode.getScopes());
-    accessToken.setExpiresAt((int) Instant.now().plusSeconds(3600).getEpochSecond());
+    accessToken.setExpiresAt(Instant.now().plusSeconds(3600).getEpochSecond());
     if (resourceServerAudience != null) {
       accessToken.setAudience(List.of(resourceServerAudience));
     }
@@ -120,7 +120,7 @@ public class SimpleAuthProvider implements OAuthAuthorizationServerProvider {
     refreshToken.setToken(refreshTokenValue);
     refreshToken.setClientId(client.getClientId());
     refreshToken.setScopes(authorizationCode.getScopes());
-    refreshToken.setExpiresAt((int) Instant.now().plusSeconds(86400).getEpochSecond());
+    refreshToken.setExpiresAt(Instant.now().plusSeconds(86400).getEpochSecond());
 
     // Store tokens
     accessTokens.put(accessTokenValue, accessToken);
@@ -161,7 +161,7 @@ public class SimpleAuthProvider implements OAuthAuthorizationServerProvider {
     accessToken.setToken(accessTokenValue);
     accessToken.setClientId(client.getClientId());
     accessToken.setScopes(scopes);
-    accessToken.setExpiresAt((int) Instant.now().plusSeconds(3600).getEpochSecond());
+    accessToken.setExpiresAt(Instant.now().plusSeconds(3600).getEpochSecond());
     if (resourceServerAudience != null) {
       accessToken.setAudience(List.of(resourceServerAudience));
     }
@@ -171,7 +171,7 @@ public class SimpleAuthProvider implements OAuthAuthorizationServerProvider {
     newRefreshToken.setToken(refreshTokenValue);
     newRefreshToken.setClientId(client.getClientId());
     newRefreshToken.setScopes(scopes);
-    newRefreshToken.setExpiresAt((int) Instant.now().plusSeconds(86400).getEpochSecond());
+    newRefreshToken.setExpiresAt(Instant.now().plusSeconds(86400).getEpochSecond());
 
     // Store tokens
     accessTokens.put(accessTokenValue, accessToken);
